added image crop functionality
