+++ src/components/FileUpload/FileUpload.css
@@ -3,3 +3,7 @@
   padding: 10px;
   box-sizing: border-box;
 }
+
+.hide{
+  display: none;
+}
